abstract shared wasm-specific code
