+++ advent_of_code/aoc/2023/wasm/Cargo.toml
@@ -12,6 +12,7 @@ repository.workspace = true
 
 [dependencies]
 aoc2023 = { path = "../solutions" }
+aoc-core-wasm ={ path = "../../core/wasm"}
 wasm-bindgen = "0.2.87"
 wasm-bindgen-futures = "0.4.37"
