+++ advent_of_code/aoc/all/solutions/Cargo.toml
@@ -11,5 +11,5 @@ readme = "README.md"
 repository.workspace = true
 
 [dependencies]
-aoc-core = { path = "../../core"}
+aoc-core = { path = "../../core/solutions"}
 aoc2023 = { path = "../../2023/solutions"}
