+++ advent_of_code/aoc/core/solutions/Cargo.toml
