+++ advent_of_code/aoc/core/solutions/src/lib.rs
